no longer spawns more enemies longer time to solve comets, more time between comets roughly doubled phillip's health
intent is for these two mooks to be a generic fight of sorts in the wilderness,
so player should have a lot of these fights without losing too much hp

too hard for me right now - played and lost 14 health. it was fun though
maybe we should have more ways to heal?